Change license
Use LGPLv3 rather than GPLv3. This makes more sense given that the
objective is for this implementation to hopefully be useful to others,
and it isn't helpful to CPS's objectives to require closed source tools
to reimplement CPS parsing (although the likelihood of such tools using
the Python implementation is probably very small).