Fix code style errors
Fix some style errors noted by pycodestyle. This is not all of them, as
some of the style rules pycodestyle wants to enforce are dumb, or would
make the code harder to read.

(Specifically, we are ignoring E261, E265, E302, and E731.)

Also, remove some cruft left over from porting cps2cmake to using
argparse (which was noted by pycodestyle due to the way it was commented
out).