Add dependency support to cps2cmake
Teach cps2cmake to emit logic to import transitive dependencies, per the
CPS package Requires attribute. This also emits a copy of the
find_dependency macro, as the one in CMake 3.8 and earlier is not
sufficient for CPS's needs.

Additionally, our internal _IMPORT_PREFIX variable has been adjusted to
include the name of the current package. This is needed so that we can
use the import prefix as part of the HINTS for finding dependencies,
without those dependencies nuking our local copy (which will happen if
the name is not unique).

See also https://gitlab.kitware.com/cmake/cmake/merge_requests/872.