Create pc2cps tool
Create a (very simple-minded!) tool to generate a CPS from a pkg-config
file. Because pkg-config is lacking in the sort of detailed information
that CPS expects and would normally provide, the resulting CPS is very
crude.

Also add descriptions of the tools to the README, and make the current
CPS version available via the main module.