Add ability to set variables earlier
It is sometimes the case that we need to change variables before
find_dependency is called; for example, because we need to manipulate
CMAKE_MODULE_PATH in order to use the find module provided by the
package to find one of its dependencies. Add logic to support new
`X-CMake-{Inclues,Variables}-Init`, which are like their existing
cousins, but are handled at the very beginning of the generated
configuration file (before finding dependencies), rather than at the
end.