Replace class variables by instance variables
In Python, if an variable is initialized in a class, but outside
a function, it will be declared as a class variable. This means that if
multiple instances of the class are instantiated, the variable will be
shared among the objects. The variables that were defined as class
variables should actually be instance variables.