Convert pkg-config link flags to CPS attributes
Teach pc2cps to try to resolve the location of the converted component's
library, and (optionally) to convert other link flags to package
dependencies. The latter is optional (and also supports a black-list) as
it often gets tripped up by the library names differing from the package
names, or by libraries that are expected to be present (e.g. m, dl) that
don't have corresponding packages. Use of this mode may well result in a
CPS that requires additional hand editing.

The `Location` attribute is resolved using `-L` paths and a prioritized
search for dynamic and static libraries. Right now, no attempt is made
to process link flags that would switch between which flavor of library
is preferred, though an option to specify this preference is provided.
The name of the library may also be specified, overriding the guess
based on the package name.

This also attempts to remove `-l` flags which reference libraries that
are covered by a `Requires`. If all `-l` flags have been removed, `-L`
flags are removed also.

Co-authored-by: Matthew Woehlke <matthew.woehlke@kitware.com>