Add cps2pc script to convert CPS to pkgconfig
The new scripts takes 2 required parameters:
* The input CPS file
* And output directory

Additionally, 2 optional parameters can be given:
* The compiler for which the pkgconfig options are created
  (default: 'gcc').
* A prefix to replace the value '@prefix@' given in the CPS file
  (default is '${pcfiledir}/../..').

A pkgconfig file will be created for each individual component listed in
the CPS file.  The pkgconfig file is generated for a specific compiler.
If not specified, 'gcc' will be used by default. This is necessary
because a certain number of parameters contained in the CPS file need to
be converted from their generic declaration to actual flags given to the
compiler. This includes definitions, link, runtime, and compile
features.

The current implementation only supports a limited number of compilers
(gcc, clang), kernels, and compiler features. This is mostly
implemented using dictionaries, which should allow for easy extension,
as necessary, by simply adding new keys to the dictionaries.

Certain CPS tags are ignored as they are not supported by pkgconfig:
* Link-Languages
* Link-Requires
* Configurations

Co-authored-by: Matthew Woehlke <matthew.woehlke@kitware.com>