Add support to cps2pc to output directly in a file
The user had to provide cps2pc a directory in which all output files
were written. Sometimes one wants to manually specify the output
file name. This is especially the case if the package contains
only one component. The user is now able to define the pkgconfig
file for that one component instead of relying on the automatic
naming scheme that is implemented in cps2pc.